Hotfix for creation of names
There is a problem with the cache and passenger. I moved the missing
function from interface utils to feature_sweeper.rb just as a hotfix so
production site can still be used. Had to change the version to rails
4.1 because production branch of subjects uses 4.1.

This is just a hotfix and needs to be addressed properly.