Fixes bug when predicate field is missing
When opening Terms the flyout tree fails due to missing field when `associated_subject_ids` has not been reindexed. This is due to the NodeMarkerPredicates expecting `associated_subjects_ids` to exist.

This shouldn't be a normal occurance, because all nodes should be
indexed, but it might happen when nodes are newly added.